move common methods to utils file
